requirements/prod.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.11.12
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
distro==1.9.0
Django==5.2.8
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework-jsonapi==8.1.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
drf_spectacular==0.29.0
et_xmlfile==2.0.0
filelock==3.20.0
fsspec==2025.10.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
groq==0.34.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ifcopenshell==0.8.3.post2
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==1.0.5
langchain-core==1.0.4
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
lark==1.3.1
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numpy==2.3.4
openai==2.7.2
openpyxl==3.1.5
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
promise==2.3
prompt_toolkit==3.0.52
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
xxhash==3.6.0
zstandard==0.25.0
psycopg[binary]
gunicorn==22.0.0
boto3==1.35.0
django-storages[s3]==1.14.3